Implement sound output for Linux via PulseAudio
I used PulseAudio because it was a lot easier to get started with than
ALSA. I'll probably switch to ALSA in the future.

In the process of making a system-specific module, I switched to GNU
Make instead of POSIX make. If I find a way to use POSIX make or
something compatible with multiple make implmentations, I'll use that.