Move sound encoding config to platform specific code
The sound encoding is not platform independent. For example, NetBSD
needs to convert to a 32-bit integer encoding before playing the sound.
Since the encoding is only used in the platform specific code, there's
no need to put it in sound.h and define macros. It's simpler to say that
prelude uses 32-bit floats to represent the sound and the platform code
can use it however it wants.