linux: Switch to ALSA
pulseaudio_simple and alsa are almost the same in this use case, so the
transition is easy. I'm not sure why I struggled with this earlier. The
ALSA docs are pretty straightforward.

In the end, the audio library doesn't make a big difference. The
interface is about the same, and the behavior is exactly the same. The
only difference is that the resulting prelude binary is 1 kB smaller,
and it links to 6 libraries instead of 26.